Move toast util to scaffold-eth folder
