+++ packages/nextjs/components/scaffold-eth/Contract/ReadOnlyFunctionForm.tsx
@@ -4,7 +4,7 @@ import { useContractRead } from "wagmi";
 import { tryToDisplay } from "./utilsDisplay";
 import InputUI from "./InputUI";
 import { getFunctionInputKey } from "./utilsContract";
-import { toast } from "~~/utils";
+import { toast } from "~~/utils/scaffold-eth";
 
 const getInitialFormState = (functionFragment: FunctionFragment) => {
   const initialForm: Record<string, any> = {};
