--- packages/nextjs/utils/index.ts
@@ -1,1 +1,0 @@
-export { default as toast } from "./toast";
