+++ packages/nextjs/utils/scaffold-eth/index.ts
@@ -1,2 +1,3 @@
 export { default as getLocalProvider } from "./getLocalProvider";
+export { default as toast } from "./toast";
 export * from "./networks";
